# service-schedule/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Service Schedule API
  description: API para gerenciamento de prestadores de serviço e horários
  version: 1.0.0
servers:
  - url: http://localhost:8080  # Altere a URL conforme necessário
    description: Servidor local
paths:
  /prestadores:
    get:
      summary: Obter todos os prestadores
      responses:
        '200':
          description: Lista de todos os prestadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceScheduleModel'
    post:
      summary: Cadastrar novo prestador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceScheduleModel'
      responses:
        '201':
          description: Prestador cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceScheduleModel'
  /prestadores/{id}:
    get:
      summary: Obter prestador por ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID do prestador
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Prestador encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceScheduleModel'
        '404':
          description: Prestador não encontrado
    delete:
      summary: Excluir prestador por ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID do prestador
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Prestador excluído com sucesso
        '404':
          description: Prestador não encontrado
  /prestadores/{id}/horarios:
    post:
      summary: Associar horários ao prestador
      parameters:
        - name: id
          in: path
          required: true
          description: ID do prestador
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Horario'
      responses:
        '200':
          description: Horários associados com sucesso
        '404':
          description: Prestador não encontrado
        '422':
          description: Horário duplicado ou inválido
  /prestadores/{id}/horarios-disponiveis:
    get:
      summary: Obter horários disponíveis do prestador
      parameters:
        - name: id
          in: path
          required: true
          description: ID do prestador
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista de horários disponíveis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Horario'
  /prestadores/{idPrestador}/horarios/{idHorario}/agendar:
    post:
      summary: Agendar horário com o prestador
      parameters:
        - name: idPrestador
          in: path
          required: true
          description: ID do prestador
          schema:
            type: integer
            format: int64
        - name: idHorario
          in: path
          required: true
          description: ID do horário
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Horário agendado com sucesso
        '404':
          description: Prestador ou horário não encontrado
        '422':
          description: Horário não disponível para agendamento
  /prestadores/{idPrestador}/horarios/{idHorario}/cancelar-agenda:
    post:
      summary: Cancelar agendamento de horário com o prestador
      parameters:
        - name: idPrestador
          in: path
          required: true
          description: ID do prestador
          schema:
            type: integer
            format: int64
        - name: idHorario
          in: path
          required: true
          description: ID do horário agendado
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Agendamento cancelado com sucesso
        '404':
          description: Prestador ou horário não encontrado
        '422':
          description: Horário já está disponível
components:
  schemas:
    ServiceScheduleModel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        categoria:
          type: string
        valorServico:
          type: number
      required:
        - nome
        - categoria
        - valorServico
    Horario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        diaSemana:
          type: string
        hora:
          type: string
        disponivel:
          type: boolean
      required:
        - diaSemana
        - hora
        - disponivel